MODINVOICE-54 Prorated adjustments - basics
Move logic to apply prorated adjustment(s) to invoice line helper (preparation for cases when invoice line affects distribution)
